ggml : move 32-bit arm compat in ggml-impl.h (#6865)
ggml-ci
